backend/requirements.txt
django>=4.2
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-filter
